Stop leadership on ZooKeeper connection lost and close connection.
Summary:
This follows the error handling suggestions in the [LeaderLatch docs](http://curator.apache.org/curator-recipes/leader-latch.html).
The leader commits suicide on ZooKeeper connection loss. It avoids closing the
latch if the client is already closed.

We could set the latch to None but this is not thread safe for now.

Test Plan: integration-test

Reviewers: aquamatthias, jasongilanfarr, timcharper

Subscribers: marathon-team, jenkins

Differential Revision: https://phabricator.mesosphere.com/D504